Move helper functions to helpers.py
